+++ solutions/rotateArray.cpp
@@ -7,7 +7,7 @@ class Solution
 public:
     void rotate(vector<int> &nums, int k)
     {
-        for (int i = 1; i <= nums.size(); i++)
+        for (size_t i = 1; i <= nums.size(); i++)
         {
             swap(nums[k % nums.size()], nums[(i + k) % nums.size()]);
         }
